+++ bh-status-list/CHANGELOG.md
@@ -8,6 +8,8 @@ Versioning](https://doc.rust-lang.org/cargo/reference/semver.html).
 
 ## [Unreleased]
 
+## [0.3.0] - 2026-04-14
+
 ### Changed
 
 - Change `kid` field in `StatusListTokenHeader` from mandatory to optional.
@@ -39,7 +41,8 @@ Versioning](https://doc.rust-lang.org/cargo/reference/semver.html).
 - Initial version of the `bh-status-list` crate.
 
 
-[Unreleased]: <https://github.com/blockhousetech/eudi-rust-core/compare/bh-status-list/v0.2.1...HEAD>
+[Unreleased]: <https://github.com/blockhousetech/eudi-rust-core/compare/bh-status-list/v0.3.0...HEAD>
+[0.3.0]: <https://github.com/blockhousetech/eudi-rust-core/releases/tag/bh-status-list/v0.3.0>
 [0.2.1]: <https://github.com/blockhousetech/eudi-rust-core/releases/tag/bh-status-list/v0.2.1>
 [0.2.0]: <https://github.com/blockhousetech/eudi-rust-core/releases/tag/bh-status-list/v0.2.0>
 [0.1.0]: <https://github.com/blockhousetech/eudi-rust-core/releases/tag/bh-status-list/v0.1.0>